Test: 지원서 - 관리자 기능 단위 테스트 추가
