+++ src/test/java/com/gdsc_knu/official_homepage/application/service/AdminApplicationServiceTest.java
@@ -19,6 +19,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
+import static com.gdsc_knu.official_homepage.application.ApplicationTestEntityFactory.createApplication;
 import static com.gdsc_knu.official_homepage.application.ApplicationTestEntityFactory.createApplicationList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -38,32 +39,12 @@ public class AdminApplicationServiceTest {
         applicationRepository.deleteAll();
     }
 
-    // TODO: 스프링부트 테스트 없이 단위테스트
-    @Test
-    @DisplayName("지원서를 열람한 경우 열람상태가 변경된다.")
-    void getApplicationDetail() {
-        // given
-        Application application = createApplication(ApplicationStatus.SAVED);
-        applicationRepository.save(application);
-        // when
-        applicationService.getApplicationDetail(application.getId());
-        // then
-        assertThat(application.isOpened()).isTrue();
-    }
-
-
-    // TODO: 스프링부트 테스트 없이 단위테스트
-    @Test
-    @DisplayName("지원서 목록 요청에 따라 필요한 데이터만 조회된다.")
-    void getApplicationsByOption() {
-
-    }
 
     @Test
     @DisplayName("메일 전송에 실패하더라도 status 변경은 저장한다.(SAVED -> APPROVED")
     void updateStatus() {
         // given
-        Application application = createApplication(ApplicationStatus.SAVED);
+        Application application = createApplication(1, Track.AI, ApplicationStatus.SAVED);
         applicationRepository.save(application);
         doThrow(CustomException.class).when(mailService).sendEach(application);
         // when
@@ -74,23 +55,9 @@ public class AdminApplicationServiceTest {
         assertThat(application.getApplicationStatus()).isEqualTo(ApplicationStatus.APPROVED);
     }
 
-    // dao 의존성 없앤 단위테스트로 변경
-    @Test
-    @DisplayName("임시저장 상태인 경우 status 를 변경할 수 없다.")
-    void failedUpdateTemporal() {
-        // given
-        Application application = createApplication(ApplicationStatus.TEMPORAL);
-        applicationRepository.save(application);
-        // when
-        applicationService.decideApplication(application.getId(), ApplicationStatus.APPROVED);
-        // then
-        assertThat(application.getApplicationStatus()).isEqualTo(ApplicationStatus.TEMPORAL);
-    }
 
-
-    // TODO: 트랙별 지원현황을 읽어오는 테스트와, 읽어온 데이터를 가공하는 테스트 분리
     @Test
-    @DisplayName("트랙별 지원서의 개수를 정상적으로 카운트한다.")
+    @DisplayName("트랙별 지원서의 개수를 정상적으로 카운트한다. (트랙에 지원이 없는 경우 0을 카운트한다.)")
     void getTrackStatistic() {
         // given
         int start = 1;
@@ -116,18 +83,6 @@ public class AdminApplicationServiceTest {
         assertThat(statistic.get("DESIGNER")).isEqualTo(0);
         assertThat(statistic.get("TOTAL")).isEqualTo(6);
         assertThat(statistic.size()).isEqualTo(6);
-    }
-
-
-
-    private Application createApplication(ApplicationStatus status) {
-        return Application.builder()
-                .studentNumber("2024000000")
-                .phoneNumber("010-0000-0000")
-                .email("test@email.com")
-                .applicationStatus(status)
-                .track(Track.BACK_END)
-                .build();
     }
 
 }
