+++ src/test/java/com/gdsc_knu/official_homepage/application/service/ApplicationServiceTest.java
@@ -1,0 +1,4 @@
+package com.gdsc_knu.official_homepage.application.service;
+
+public class ApplicationServiceTest {
+}
